pyproject.toml
[tool.poetry]
name = "churn_prediction"
version = "0.1.0"
description = ""
authors = ["Alphonse_Brandon"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.0"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
kaggle = "^1.5.16"
ipykernel = "^6.25.2"
pandas = "^2.1.1"
seaborn = "^0.12.2"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
